moved healpix functions; added Map.healpix()
